+++ src/main/java/com/doublew2w/sbs/mybatis/builder/BaseBuilder.java
@@ -1,6 +1,8 @@
 package com.doublew2w.sbs.mybatis.builder;
 
 import com.doublew2w.sbs.mybatis.session.Configuration;
+import com.doublew2w.sbs.mybatis.type.TypeAliasRegistry;
+import lombok.Getter;
 
 /**
  * 建造者模式：建造者基类
@@ -11,9 +13,11 @@ import com.doublew2w.sbs.mybatis.session.Configuration;
  */
 public abstract class BaseBuilder {
   protected final Configuration configuration;
+  protected final TypeAliasRegistry typeAliasRegistry;
 
   public BaseBuilder(Configuration configuration) {
     this.configuration = configuration;
+    this.typeAliasRegistry = this.configuration.getTypeAliasRegistry();
   }
 
   public Configuration getConfiguration() {