+++ src/main/java/com/doublew2w/sbs/mybatis/mapping/package-info.java
@@ -1,0 +1,2 @@
+/** 映射的基础包 */
+package com.doublew2w.sbs.mybatis.mapping;