+++ src/main/java/com/doublew2w/sbs/mybatis/transaction/jdbc/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * JDBC事务
+ */
+package com.doublew2w.sbs.mybatis.transaction.jdbc;