+++ src/main/java/com/doublew2w/sbs/mybatis/transaction/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 事务的基础包
+ */
+package com.doublew2w.sbs.mybatis.transaction;