+++ src/main/java/com/doublew2w/sbs/mybatis/type/package-info.java
@@ -1,0 +1,2 @@
+/** 存放各种类型 */
+package com.doublew2w.sbs.mybatis.type;